Setting State Correctly
Using optional syntax. In nameChangeHandler added anonimous arrow func to setState(), which receives prevState and props as arguments, and in the anonimous function body we need to return new state object. By using prevState React guarantee that prevState would be as we expect it to be. Best practice for state updates, that depends on the old state !!!