Using Context API
Created context folder, and added auth-context.js file. In file we create Context object. Assign to a authContext constant a React.createContext() and initiliaze the default value for our Context.

Context is designed to share data that can be considered “global” for a tree of React components, such as the current authenticated user, theme, or preferred language.